AES_GCM
The AES_GCM aimed to match what was shown in the picture. I made it a class that takes the first 32 bytes of the key. It can change it if needed. It encrypts data and gives out the iv, ciphertext, and tag. It can use those outputs to decrypt the message. I didn't remember how to add the library to github or how to make the venv. I did find online how to add a venv but I didn't put it in this one in case it doesn't work.